Scrub sensitive info in map
Publish context is being logged which now contains sensitive pass-phrase information. This PR would scrub it before writing it to log.

This change was somehow missed in the last PR on which csi-driver's PR https://github.com/hpe-storage/csi-driver/pull/246 is dependent on. @raunakkumar - request you to review it following which I'll update the go.mod in csi-driver PR.